Render some stuff on the server (#435)
Nothing crazy, but making it easier to add sever-side rendered stuff in the future. Not everything needs to go through react and actually it's sorta dissuading to have to write so much boilerplate for very simple features. This should allow me to add simple things in the future without having to deal with typescript and babel and graphql and caching.