Move exenv to twitch app
Not being used anywhere else and tricky to manage for an umbrella app,
so just moving it to the place where it's needed for now.

Since credentials are stored in memory belonging to the Exenv
application, it's necessary to start that app before credentials are
accessed. Previously the app was under supervision of the
`client` app, which was fine in dev since everything is running there.
But as only some apps are started in test mode, it became annoying to,
for exmaple, start a process in the `client` app when we were only
testing the `twitch` one. For this reason it's probably best to specify
`exenv` as a dependency for each app that needs it and maintain separate
credentials for each.