Remove user from heroku redis url (#518)
Redis doesn't actually support a `user` and our redis library `redix`
recently started raising an error when it encountered this param in the
URL (previously it would just ignore it). This is kinda annoying on
Heroku, though, since it always includes a `h` user by default.

This PR grabs the heroku url, removes the `h:` portion, and then passes
it along to redix.