Rename Twitch.Repo -> Twitch.GoogleRepo (#517)
Going to migrate away from using Google SQL, to make things a little
nicer it'll be easier to maintain connections to both databases for a
time. To that end we'll first rename Twitch.Repo to Twitch.GoogleRepo,
add a new Twitch.Repo (which should be the new thing everything uses),
and then eventually migrate everything to use Twitch.Repo again.