Add a new Twitch.Repo (#539)
The old one was renamed to `Twitch.GoogleRepo` since it uses Google SQL.
We're trying to migrate away from that, so this adds a new `Twitch.Repo`
back, but this time it points to heroku postgres. This way we'll have
connections to both old and new databases. This should make migration
not too bad.