Cache get requests to twitch
Since we need to make a lot of these in quick succession, we can speed
things along quite a bit if we use some rudimentary caching logic.

This works by generating a cache key based on the path, params, and
headers of a request, and then storing the response in-memory for future
retrieval.

An additional improvement would be to probably only cache 2xx responses,
since as-is this will cache everything (even 5xx's)