Allow subscribing to chat anonymously
Previously the channel route would only display chat events if someone
had already created a subscription to that channel in the react app.
This is kinda not ideal, so instead let's create a subscription (just to
chat--no emotes) if someone visits a channel route for which there isn't
an active subscription.