Handle unknown emotes
Sometimes TwitchEmotes takes a little while to update its database when
new emotes are added. This results in a gap between Twitch supports an
emote and when the TwitchEmotes service allows querying for it. During
this window, Twitch chat events will tag these emotes with their ID, but
querying TwitchEmotes for this ID won't return anything.

Previously we always expected TwitchEmotes to return all emotes tagged
by Twitch, which resulted in a lot of errors during the aforementioned
window. This PR just ignores emotes which TwitchEmotes doesn't know
about, since we need a response from it to do anything else with the
emote.