Delete unconfirmed webhook subscriptions
When initially creating a webhook subscription with twitch, twitch will
make a GET request to the callback URL confirming that it's received
your webhook and everything is working.

This PR waits 30 seconds after creating a subscription, checks to see if
it was confirmed by twitch, and deletes it if it wasn't. This way we
don't end up with a bunch of random subscriptions laying around if
something goes wrong.