Use raw body for calculating hmac hash
Because of some pedantic architeture decision, it's only possible to
read the raw request body a single time, after which it will always be
`nil`. Since we're using some Plug helpers which read and parse the body
before we have access to it, there's no way for us to actually see what
is. This PR adds a custom plug to save the raw body in the conn's
assigns so we can look at it later on. Kinda annoying that this is the
only approach here, but here you go.