Persist webhook callbacks
Now that we know what callbacks look like, we can begin persisting them.

We might also wanna attach the `stream_id` in the callback record so
it's easier to know what stream was live at the time the callback was
received. This should make querying for the video easier later on, I
think.

If we go down that path, though, we should support a `cache: false`
option or something to the stream query, since having a stale
`stream_id` wouldn't be idea.