Calculate batch ingredient cost correctly
Also save it to the database so we don't have to calculate it every
time.

In order to update existing batch ingredients, I had to add a primary
key to the `soap_batch_ingredients` table. This would have been
necessary anyway to support updating batch ingredients, so I figured why
not just go for it here.