bump erlang version on heroku (#1143)
The elixir buildpack doesn't have the exact same version we use in dev,
so use the closest one available (off by just a single patch version)

We ran out of memory when building on heroku. Try to ask node to use
only 256mb of memory. This might not be enough...