Upgrade phoenix_live_view to 0.8
When I started this, I was getting an error saying that a module called
`ClientWeb.Router` couldn't be found. Some initial research pointed me
in the direction of a circular dependency on the router somewhere.
For example, this can happen when a plug imports anything from a router,
since the router needs the plug to build before it's compiled, and the
plug needs the router to build. Since these types of things are pretty
difficult to check for at the compiler level, there's not much that
Elixir can do except say it couldn't find a compilation dependency and
exit.

Apparently a lot of this compiler ambiguity can be resolved by more
explicitly referencing the functions of the router, instead of just
importing the thing wholesale. Still not knowing what the true cause
was, I went about enacting this. This meant that I changed `import
Router.Helpers` to `alias Router.Helpers, as: Routes`, and updated all
route references with a `Routes.` prefix. Once I did this--it took a
little bit--I realized that the true issue was that I was referencing a
plug in the router which didn't exist anymore (it was removed in
live_view 0.8).

I tried backing out all of my `Routes.` changes and just removing the
plug in question. Turns out that this change was sufficient, so it
really isn't necessary to change all routes in the entire app. Still,
because it made it so much easier to debug the actual issue here, I
think I'll leave them in place to make life easier on my future self. I
do also prefer being more explicit about references instead of relying
on imports, anyway, so this seems like a better solution regardless.