Fix path to config files (#1176)
* Fix path to config files

This broke a deploy on heroku (but not circle?)

* Combine config files

There seems to be a little bit of consensus about this being the correct
way of handling config for umbrella apps. While I think I'd prefer
organizationally to keep things separate, keeping it all together does
dispell the notion that these apps can be configured independently,
which I think is good. Before I don't think it was super obvious that
this was the case. It also wasn't really clear what the root-level
config files were for.