Improve CI caching in build_prod step (#1208)
Similar to what we did for test and dev in #1205, also cache `deps` in
the prod cache. I just missed this in the last PR.

Also touch build files as a workaround to make rebar3 not recompile
erlang files. Currently is uses the timestamp, but that gets weird when
we restore a cache. They're apparently working on using a hash, but that's
not ready quite yet.