Add /one-time-login-link endpiont
This allows a user with a valid API token to generate a one-time login
link, which authorizes a session with the frontend. By default it will
just redirect you to the landing page, but you can also specify a `to`
parameter to be redirected to any arbitrary path.

This can be used with the water logs kiosk to automatically log a user
into the app.