Make liveview test slightly more stable
I could not figure out how to make this test wait for rendering to
complete like all the others do. It's because when an update occurs, we
receive a Phoenix.PubSub event, and then use `send_update` to get child
components to re-render with updated counts. Getting the children to
update takes a little while, though, and sometimes the assertion happens
before the rendering is done, and results in a failure. LiveView has
helpers that are supposed to prevent this from happening, but I guess
they do not work in this particular case.