Fix generated date series
It was generating dates with timestamps, so the query would only match
an entry if it was added at 00:00:00. In the tests this is actually what
we were doing, so it resulted in some false positives earlier on.

To get `generate_series` to generate a series of dates without
timestamps, we have to explicitly cast the input to this type.