Self hosting (#2295)
Migrate away from Heroku, which has felt sorta unmaintained for a while, and is now starting to charge more for personal projects. Instead I'll just host it myself. 

To do this, we switch to using Elixir releases to build and bundle our app, which is then run on a server I have using systemd. Right now this deploy process all relies on being on the same network as the server, but I can improve that in the future if it becomes a pain point.

Because I'm running on an M1 Mac and my server is running on Linux, I can't just run `mix release` myself--it'll generate a binary for the wrong processor architecture. Instead, this uses Earthly, a sorta platform-agnostic CI runner, to build the release in a docker container. The release is then manually copied over to the server and then started. Maybe in the future I'll build a docker image or something, but that requires using a docker registry and all that, which is a bit of overhead I'm not necessarily keen to take on atm. 

Right now, Earthly is running on the same server that the app is eventually started on. But the nice thing about Earthly is that it could run anywhere. So if next week I decide to run it on GHA, I could do that without making much change to my Earthly config. The thing then is where to put the release artifact, and how to handle secrets.