Logging out successfully redirects
Previously, after you logged out, you'd still just stay on whatever page
you were on. When you performed another action, we'd detect you weren't
logged in, and redirect you to login. This is because we didn't update
the cached login state, so the client didn't know anything had changed.

This fixes that by running another query to update the cache after we
logout.