Bump esbuild to v0.15
This takes us from the last release of the 0.14 series to the last
release of the 0.15 series. 0.17 is the latest, but I don't necessarily
want to do that all in one jump. Something about our previous
dependencies kept us limited to 0.14 for a while--probably semantic--and
now that we're free we can get back up to latest.