Add cost to ijust events (#2529)
You can now edit an ijust event to modify its name and a new `cost` field. The cost field will eventually be used to calculate the per-use cost of an item, but for now we're just recording the cost (this PR was already big enough!) 

Also adds an `absinthe_error_payload` package which helps with converting changeset errors into graphql errors. This follows the errors-as-data approach, which is nice. This was the first time we needed to return validation errors, which is pretty wild considering how much graphql is used in this app.

Something I'd like to explore in the future is a more lightweight way of adding forms in react. It's just an absolutely huge overhead to add all the stuff for `yup`, `react-hook-form`, and typescript. This cannot be the best way of doing this.