Remove Auth application (#2535)
I had originally split this into a separate app for some reason. I was
new to umbrella apps at the time and was trying things out, I guess. Now
I realize that it's simpler to just put everything in the Client app,
especially since there's no need to share this with anything else.