Rotate encryption secret
The encryption secret (stored at `key.txt` and used to encrypt our
environment vars) was in need of rotating, so here I replace it with a
new one. I also switched a few of the scripts here to use the `op` cli
rather than requiring you to open up 1password to get the secret.