Remote twitch emote API (#2828)
Previously we used [twitchemotes.com](https://twitchemotes.com/) to fetch data about each twitch emote as it was used. This stopped working at some point, probably because they changed their (undocumented) API somehow.

Turns out we actually don't need to use that service anyway, since the only info we need is the ID and the code, both of which are present in the chat message. So in this PR I remove that and update the chat parser to extract twitch emotes automatically. 

This should address one of the more common sentry errors we're seeing and hopefully do a better job of keeping us well beneath our error limit there.