Better error handling for Awair
Allow each Monitor process to stay alive even if it's not connected.

Previously the Monitor exited if it wasn't able to connect to the Awair.
The supervisor would retry the process forever. This works fine but
results in a lot of errors, since each Awair is intermittently
unavailable.

Now the Monitor will handle the failed connection and remain alive,
retrying periodically with the same frequency as it would check in
normally. This feels like a better approach for a few reasons:

* It will reduce errors to Sentry
* Awair unavailability is a somewhat expected state so the genserver
  should be okay with that from time to time
* It will allow us to retry in a less spammy way. The supervisor will
  immediately resart the process, but that just results in another
  failed attempt. In reality we just need to wait a minute or so and the
  Awair will become available again, so there's no need to make multiple
  requests in such quick succession.

Also add .projecgtions.json to make it easier to switch between test and
implementation files