Fix issue with downloading archived release (#3123)
I guess this is becuase I switched to using a new version of the
upload-archive action, which now uses some different service for
storing the files. I guess the issue was that `wget` forwards the
Authorization header when following redirects. That I guess messes up
one of the many redirects that we end up following when trying to
download an archive. I was able to fix the issue by just using curl, as
the gh docs recommend. Probably I could have gotten wget to work but it
was easier to just copy and paste this example from the api docs.