Submit is enabled by default
Keeping it disabled until the form is valid is ok but requires some
additional logic that I don't feel like doing right now. Since the form
is only valided when an input is blurred, we can make that more seamless
by having the submit button being always enabled, since when a user
presses enter or cliks the button it will blur the input and validate
the form.