Automerge some dependabot updates (#3276)
Some of these (mostly @types/) deps are updated frequently but never
require human intervention to merge. So long as CI passes it's fine to
merge these automatically.

This sets up a new action that does that, starting with just
`@types/node`, and just for patch and minor updates. I may expand that
to other deps in the future.